+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 2025-07-03
+
 ## Added
 - community: add rainbow-delimiters with theme support
 - community: add color picker ccc-nvim
@@ -12,6 +14,7 @@
 - practicalli: revert to upstream other.nvim
 - practicalli: remove parpar in favor of clojure pack
 - dev: ci scheduled stale issue & pr check (monthly)
+- practicalli: snacks notifier wrap text
 
 # 2025-04-24